+++ flask_dance/__init__.py
@@ -1,4 +1,4 @@
 "Doing the OAuth dance with style using Flask, requests, and oauthlib"
 from .consumer import OAuth1ConsumerBlueprint, OAuth2ConsumerBlueprint
 
-__version__ = "6.2.0"
+__version__ = "7.0.0"