+++ CHANGELOG.rst
@@ -3,6 +3,11 @@ Changelog
 
 `unreleased`_
 -------------
+nothing yet
+
+`7.0.0`_ (2023-05-10)
+---------------------
+* Removed Twitter pre-set configuration
 * Added Dexcom pre-set configuration
 * Added support for authorization flow with PKCE_
 
@@ -398,7 +403,8 @@ Fixed
 .. _flit: https://flit.pypa.io/
 .. _PKCE: https://www.rfc-editor.org/rfc/rfc7636
 
-.. _unreleased: https://github.com/singingwolfboy/flask-dance/compare/v6.2.0...HEAD
+.. _unreleased: https://github.com/singingwolfboy/flask-dance/compare/v7.0.0...HEAD
+.. _7.0.0: https://github.com/singingwolfboy/flask-dance/compare/v6.2.0...v7.0.0
 .. _6.2.0: https://github.com/singingwolfboy/flask-dance/compare/v6.1.1...v6.2.0
 .. _6.1.1: https://github.com/singingwolfboy/flask-dance/compare/v6.1.0...v6.1.1
 .. _6.1.0: https://github.com/singingwolfboy/flask-dance/compare/v6.0.0...v6.1.0